0811 배경인 - 계좌 > 포인트 충전 수정중

+++ src/main/java/com/pj2/shoecream/vo/AdminAccountInoutVO.java
@@ -4,6 +4,11 @@ import java.sql.Timestamp;
 
 import lombok.Data;
 
+
+// tran_type : 입금인지 출금인지 확인 코드
+// d : deposit (입금)
+// w : withdraw (출금)
+
 @Data
 public class AdminAccountInoutVO {
 	   private int admin_account_idx;
@@ -11,5 +16,5 @@ public class AdminAccountInoutVO {
 	   private int tran_amount;
 	   private String tran_type;
 	   private Timestamp tran_date;
-	   private int admin_account_balance;
+	   private Long admin_account_balance;
 }
